allow for delete on form, re #308
