+++ src/app/editor/components/ResourceEditPage.vue
@@ -60,7 +60,7 @@
             </div>
         </v-ons-list>
         <div v-show="showForm">
-            <resource-edit-form :back="back" :tile="tile" :tiles="tiles" :card="card" :save="saveTile" />
+            <resource-edit-form :back="back" :tile="tile" :tiles="tiles" :card="card" />
         </div>
     </v-ons-page>
 </template>
@@ -298,52 +298,6 @@ export default {
                 return this.getCardinality(card) === 'n' || this.hasTiles(card) === false;
             }
             return false;
-        },
-        canDelete: function(tile) {
-            return tile.tileid && tile.tileid in this.$store.getters.activeProject.newly_created_tiles;
-        },
-        getCardinality: function(card) {
-            var found = this.$underscore.find(this.allNodegroups, function(nodegroup) {
-                return nodegroup.nodegroupid === card.nodegroup_id;
-            }, this);
-            if (!!found) {
-                return found.cardinality;
-            }
-            return '1';
-        },
-        saveTile: function(tile) {
-            console.log('saving...');
-            console.log(tile);
-            this.$emit('saving', true);
-            var self = this;
-
-            this.$store.dispatch('persistTile', tile)
-                .then(function(savedTile) {
-                    return savedTile;
-                })
-                .finally(function() {
-                    console.log('tile save finished...');
-                    window.setTimeout(function() {
-                        self.$emit('saving', false);
-                    }, 2000);
-                });
-        },
-        deleteTile: function(tile, e) {
-            console.log('in deleteTile');
-            console.log('tile: ', tile)
-            var self = this;
-            e.stopPropagation();
-            this.$ons.notification.confirm({
-                message:  'Delete this Data? This can\'t be undone.',
-                callback: function(answer){
-                    if (!!answer) {
-                        self.$store.dispatch('deleteTile', tile)
-                        .finally(function() {
-                            console.log('tile delete finished...');
-                        });
-                    }
-                }
-            });
         }
     }
 };
